Add display field for Users to show user avatar (WiP)
