+++ frontend/src/app/modules/fields/display/field-types/wp-display-user-field.modules.ts
@@ -26,33 +26,25 @@
 // See doc/COPYRIGHT.rdoc for more details.
 // ++
 
-import {ResourcesDisplayField} from "./wp-display-resources-field.module";
-import {cssClassCustomOption} from "core-app/modules/fields/display/display-field.module";
+import {DisplayField} from "core-app/modules/fields/display/display-field.module";
 
-export class MultipleLinesStringObjectsDisplayField extends ResourcesDisplayField {
-
-  public render(element:HTMLElement, displayText:string):void {
-    const values = this.value;
-    element.setAttribute('title', displayText);
-    element.textContent = displayText;
-
-    element.innerHTML = '';
-
-    if (values.length === 0) {
-      this.renderEmpty(element);
-    } else {
-      this.renderValues(values, element);
+export class UserDisplayField extends DisplayField {
+  public get value() {
+    if (this.schema) {
+      return this.attribute && this.attribute.name;
+    }
+    else {
+      return null;
     }
   }
 
-  protected renderValues(values:string[], element:HTMLElement) {
-    values.forEach((value) => {
-      const div = document.createElement('div');
-      div.classList.add(cssClassCustomOption, '-multiple-lines');
-      div.setAttribute('title', value);
-      div.textContent = value;
-
-      element.appendChild(div);
-    });
+  public render(element:HTMLElement, displayText:string): void {
+    if (this.placeholder === displayText) {
+      this.renderEmpty(element);
+    } else {
+      element.innerHTML = `
+        <div class="avatar-mini avatar-default">${displayText.charAt(0)}</div>
+      `;
+    }
   }
 }
